back/contracts_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.9.0
pydantic==2.5.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.27.0
